Search thread
Spawn search in it's own thread.  When another search is started tell
the thread to quit the start the next search.  Every 1000 nodes we will
check to see if we got the signal to quit.  Code is a bit jank but it
seems to work and we can interrupt thinking when we give the engine a
new position.