Transposition tables
This adds transposition tables to search.  Modify search to cache node
results and to use them when possible.  We need to modify the search to
use a score enum instead of an integer score, because of the caching.

Add a test for mate in 2 to make sure things work.  To test this code
manually, I loaded a mate in 5 position and made sure the engine was
able to execute it consistently.  In buggy implementations, I would
sometimes see cases where the engine would find mate in 4, then fail to
see mate in 3 the next turn.