Iterative deepening with hash move
Provide the hash move to the search function and search that move first.
This lets us prune the results better because it will raise alpha
immediately to a higher value.  To further improve this we implement
iterative deepening.  When doing search, instead of doing depth D
immediately we search depth 1, 2, ..., D.  This populates the
transposition table with partial results.  Overall the algorithm ends up
being significantly faster than the previous version.