Spend less time in malloc
We were spending a lot of time in malloc.  Allocating a Vec with
capacity 128 instead of default significantly reduces the number of
reallocs, improving perft by quite a bit.  We also remove map/filter in
quiescence search by only returning captures from movegen, and doing
sort directly on the Vec without creating a new one, and that saves some
allocs as well.  Allocs are now taking 6.1% of the time in search, and
dealloc 5.5%.